requirements.txt
flask>=3.0.0
pyusb>=1.2.0
Pillow>=10.0.0
